SWC-6318 - Move test files to /src/__tests__
