+++ src/__tests__/lib/containers/entity/page/title_bar/TitleBarProperties.test.tsx
