+++ src/__tests__/lib/containers/favorites/FavoriteButton.test.tsx
